Refactor UserAnswers for form display.
Added methods to UserAnswer, like `UserAnswer.single_answer_types`. A bit more
specific than the previous `UserAnswer.answer_types` (which still exists) was.

In UsersController, `@single_answers` was using `where` to get the UserAnswers
that had been built in the `all_demos.each` block. That was my bad -- a silly
mistake -- because `where` returns the result of a SQL query, which means built
objects that aren't in the database won't be part of the returned collection.

So instead I'm using `@user.user_answers.map` to collect UserAnswers that are
single answer questions. (And `compact` removes any nil values from the array.)